Inject profile email into enterprise url lookup service
Remove a dependency on Profile by injecting a callback computing the
profile email.

Using IdentityManager directly is not an option because
OIDC-managed profiles might not have a signed in identity. In this
case the email is stored in the profile prefs.

OIDC-managed profiles are not available on iOS. The iOS factory will
fetch the email via the IdentityManager.

This is part of the effort to move the enterprise service to components
so it can be used for iOS Url filtering.

Bug: 399376916
Change-Id: I7e213e1a602d4be2a56016a83dc954cc93b3b13c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6388396
Reviewed-by: thefrog <thefrog@chromium.org>
Commit-Queue: Ernesto Izquierdo Clua <eic@google.com>
Cr-Commit-Position: refs/heads/main@{#1437517}