Make DeclarativeNetRequestBrowserTests cover PA TKVv2 requests.
We're adding caching logic to Protected Audiences trusted key value
version 2 requests, which use a very different flow from the
non-cache path. This CL enhances and renames a pre-existing
DeclarativeNetRequestBrowserTest to make sure TKVv2 requests are
wired up to extensions.

This CL moves privacy_sandbox_coordinator_test_util.h, needed to set
up a coordinator to enable these fetches, to content/public/test/,
and makes InterestGroupManagerImpl::AddTrustedServerKeysDebugOverride()
a member of InterestGroupManager.

It splits DeclarativeNetRequestBrowserTest.FledgeAuctionScripts in
two, since leaving it as a single test resulted in unexpected failures
that seem unrelated to the changes to the test - an extension
seems to have stopped redirecting requests not modified by this CL
for some reason. The test was not flaky before, but flaked twice
on Android before being split up.

Bug: 330322161
Change-Id: If3fd113cce8ed6ef7dae7853312cb8f3a5be3d43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6330419
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Reviewed-by: Kevin Graney <kmg@google.com>
Cr-Commit-Position: refs/heads/main@{#1437524}