Update disclaimer text for history embeddings
If policy is set to not allow logging, change the disclaimer under the
search input and inside the promo to a different string. Since the text
is long now, make the width of the disclaimer smaller so that more of it
wraps and set text-wrap to `pretty` so the browser can wrap accordingly.

http://screen/E8o8ki3zeYumvLT

Fixed: 374323226
Change-Id: Ied404074590e7f9fa05a3da7649479353b4a4f4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6385243
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437562}