Add a feature to disable software GL fallback due to crashes.
The AllowSoftwareGLFallbackDueToCrashes feature controls if the GPU
process will fall back to software GL (SwiftShader or WARP) due to
repeated crashes.

The SOFTWARE_GL mode is still added to the list of
fallback modes because it is part of the normal flow for blocklisted
GPUs.

Bug: 402494900
Change-Id: I6f178d612b98874d3469b1a2a083b78bc1338429
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6362999
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437571}