[ios] Avoid retaining SceneState in DefaultBrowserBannerAppAgent
The DefaultBrowserBannerAppAgent was retaining the SceneStates via
the map of SceneState* to the cached state.

Change the std::map<...> to an NSMapTable. Since the isUIEnabled
and isIncognitoContentVisible value do not need to be cached (as
SceneState only notify its observer when the value change), the
only state that needs to be cached is whether the SceneState was
in foreground or not.

Add helper method to store and retrieve the cached value and hide
the access to the NSMapTable.

Bug: none
Change-Id: I078fe846c287b28ef08529e5f0bfb58ebdded89d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387226
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1437583}