[base/feature] Improve BASE_FEATURE_PARAM() documentation
If a server-side config provides a feature param value for a disabled
feature, then this is not reflected on the client side. This was
documented already, but not in the most obvious place, now that
FeatureParam are declared with BASE_FEATURE_PARAM(). Add a warning there
too.

Change-Id: I6e8715f06faf5adf7f78969476f24651a41f29ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387189
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Auto-Submit: Benoit Lize <lizeb@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437585}