Extend Autofill Ai private API to support dates
This information will be used by the Autofill Ai settings page to add
<select> elements for dates.

Selectors in date fields will roughly look like this:
https://screenshot.googleplex.com/AxQNBNLxgceLZPb

3 selectors, one for month, one for days, one for years.

The selectors can all 3 be left completely empty (this would result in
3 empty strings in the UI), or all 3 populated.
Incomplete dates, where only some of the selectors are populated, are
considered invalid.

If the 3 selectors are all empty, and the user clicks "Save", then the
respective attribute instance for that date field is not passed to the
private API.
This matches the behavior of input fields. If an input field is empty,
and the user clicks "Save", then the respective attribute instance for
that field is not passed to the private API.

Validation of dates is treated differently because they are the only
field that can invalidate the add/edit dialog by being only partially
complete.

Bug: 393318914
Change-Id: I54f5c440b398d8bf28fca16e7de0ca787f47a004
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6362140
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Tim <tjudkins@chromium.org>
Commit-Queue: Theo Cristea <theocristea@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1437586}