[Combo button] Hook up button to bubble host
Creates the concept of a 'permanent' pinned toolbar button which
continues to exist when not visible. This allows it to be linked to tab
search bubble host, which expects a consistent button reference. As of
this CL, the menu item and toolbar button both function as expected.

Bug: 404319632
Change-Id: Ie4b6d412bc670935ee8eb0e0c41a9e7d8f5fbbf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387361
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Emily Shack <emshack@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437628}