Fix flaky test testSaveStateNoFlatBufferPrior
The legacy Tabstate file is now deleted
asynchronously as it is no longer needed.
This needs to be periodically checked for in
the tests until the delete occurs as there
is no synchronization. There is currently
a race condition between the test and
the deletion.

Bug: 406258165
Change-Id: I41c355d3a6b147e4c01b8f584188ca70cce03cff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6393035
Reviewed-by: Brandon Wylie <wylieb@google.com>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437664}