Fix double padding in dialogs by explicitly setting decorFitsSystemWindows
This change fixes the double padding that was being applied to dialogs when drawing edge-to-edge. It seems like, on some devices, the window containing ComponentDialogs would not be drawing edge-to-edge unless Window#setDecorFitsSystemWindows(false) was explicitly called, regardless of the signal given by the Window's DecorView's #getFitsSystemWindows(). Making the explicit call does seem to fix things, by preventing the duplicate padding that was being applied to some of the root views in the ancestry of android.R.id.content.

Bug: 403383825, 404184778
Change-Id: I22884869b68d7b61e3a842f923cc28f8ec76aaa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6389535
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Commit-Queue: Charles Hager <clhager@google.com>
Cr-Commit-Position: refs/heads/main@{#1437693}