[TabRestore] Represent single tab windows as window entries in the UI
Windows that are closed with a single tab are now represented as
window entries with a single tab in the recent tabs submenu.

This CL removes the logic from the service meaning other platforms
using it the service are not subject to desktop specific logic.

Removes SessionRestoreTest.WindowWithOneTab since it is no longer
applicable.

Updates remaining tests to account for this change.

Change-Id: I0168dbadfecd4e9b611b91989180e3900d40627b
Bug: 41227458, 40438075, 40846749
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387072
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437715}