[iOS] Reduce the size of the KA gradient
The Keyboard Accessory's suggestion chips gradient was visibly too
large on tablets. This CL reduces the size of the gradient a bit
on phones and significantly on tablets to make it more in line with
the expected visual effect.

Bug: 385172448
Change-Id: I9fc4a3b44e73a26b5bc29c4eacfebefc05baf58c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6388700
Auto-Submit: Alexis Hétu <sugoi@chromium.org>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437735}