pdf-default: add GetChromeDefaultFileHandlerState
This change adds a shell_util method to check whether Chrome is the
default handler for a file extension. In order to reuse the existing
code in ProbeCurrentDefaultHandlers, it removes ProbeProtocolHandlers
and has callers instead pass an association type (protocol or file
extension) to indicate what they want to query. It also adds a helper
function to pass a single protocol/file extension, since there are now
two places that need to do that.

Unfortunately this code isn't very testable since it culminates in a
call to a Windows API. The existing code also has no tests.

Bug: 396202897
Change-Id: I6220462761ff46f3dd05c6069a5ffd60e158440a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6357734
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/main@{#1437752}