Replace StrictMock with NiceMock in ClientSideDetectionHostTest
I don't see a strong reason that we need to enforce a strict behavior
for the underlying classes. We'd like to enforce the expected calls
of these classes, but we are uninterested in the calls during setup.
StrictMock makes the test brittle and complicate the set up process.

Reference: go/gmockcook#NiceStrictNaggy (sorry I can't find a public
version of it).

Bug: 405452036
Change-Id: Ie042879d7de3eabe32ad29ca6cd0e30b0728b1a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6370023
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Seung Jae Lim <andysjlim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437758}