AXTreeFixing: Makes several fixes to the way that observers are created
This patch takes the opportunity to: 1. Fix the WebContentsObserver list
by switching to an std::vector. It turns out that base::ObserverList is
used for notifying observers, not for observing something. 2. Simplify
the AXMode comparison. 3. Prevent memory leaks of observers by using a
unique_ptr. 4. Check if the WebContents have been destroyed before
attaching an observer. 5. Add / remove observers when the AXMode for the
current process plus profile has been turned on / off because
`OnAXModeAdded` is not called when individual web contents have had
their AXMode changed. 6. Move some computations outside the observers
loop for efficiency due to their repeated nature. 7. Made a virtual
function pure as it should be.

A followup patch will also watch for the creation and destruction of new
`WebContents`.

AX-Relnotes: n/a.
Bug: 399384017
Change-Id: Id61dd5556eef94fc8ee8cf0cea6413a48668a2d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6393856
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1437771}