[PWA] Fix flakiness in focus existing tests for navigation capturing
This CL fixes flakiness in NavigateBrowserUsedForFocusExistingAppWindow
by removing the wait for browser activation, and introducing a waiting
logic to wait for the launch params to be enqueued in the test site.

This is done by introducing a launch_detector script that adds itself as
a consumer to window.launchQueue in the test sites, and an existing
utility function that obtains the launch parameters that have been
enqueued in the window using content::EvalJS.

Even though WebAppNavigationCapturingBrowserTestBase had that function,
the test fixture isn't available on CrOS yet, which is why it cannot be
used inside NavigationCapturingBrowserNavigatorBrowserTest. The utility
function is thus moved to a central location closer to other test utils.

Bug: 402284172
Change-Id: I282c0bb0956c593c38124da9fca4b8f472083be5
Include-Ci-Only-Tests: chromium.mac:mac14-tests|browser_tests
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6389136
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437780}