Sunfish: Migrate Lens Web functions to delegate
Migrate most of the Lens Web-related functions back to the chrome/
delegate, so that most of the operations are hidden to the capture mode
controller and it becomes easier to create a mock panel for testing.
This CL is part of an effort to enable the Lens Web API implementation
by default.

Bug: 398259275
Test: SunfishTest.*
Change-Id: I8b0a5cf70bb6b87e44d85073410062967c9f6945
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6363315
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Elijah Hewer <hewer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437786}