Propagate static position's align-self direction
Propagate the correct 'align-slef' direction added to
LogicalStaticPosition and PhysicalStaticPosition in CL:6383615
up the tree.

This involved adding the `align_self_direction` to
PhysicalOofPositionedNode and the fragmentation variant. This also
involved adding the `align_self_direction` to
`AddOutOfFlowChildCandidate()` within the FragmentBuilder,
defaulting to kBlock.

The conversion of this direction was added and tested in CL:6383615.

The align direction will be used in OOF layout to properly handle
alignment when an OOF is statically positioned.

Bug: 40206357
Change-Id: Icfed7cf3a0c9d67d1300b528ae2093802f0caa0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6383756
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1437814}