[Media] Stop resetting WebMediaPlayerImpl visibility state on pause
Every time that WebMediaPlayerImpl::Pause is called, the private
variable visibility_pause_reason_ gets reset. This leads to bugs where
media elements in a hidden tab (backgrounded), which can be paused for
optimization reasons, are not properly resumed.

For example, when switching to a new tab, a backgrounded muted video
element can be paused by the browser for optimization. Then, when we go
back to the video's tab, the video should be resumed automatically.
However, if WebMediaPlayerImpl::visibility_pause_reason_ gets reset
while the video's tab is hidden, the video may not always resume.
Currently, this can happen if the HTMLMediaElement seeks while its tab
is hidden.

This CL refactors the WebMediaPlayer::Pause method to receive an
std::optional<PauseReason>. Depending on the PauseReason value, may be
reset or not. New tests for the have been implemented in
web_media_player_impl_unittest.cc to test bug scenarios.

Bug: 403417910, 351354996
Change-Id: I9e19595ed8b89259acaeb5d533474d082c0beab6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6369530
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Gabriel Brito <gabrielbrito@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1437856}