Support partitioning for target="_blank" links
When a link click opens the page in a new tab, the previous primary main
frame and opener values are severed. Prior to this CL, this meant that a
triple-partitioned :visited links key could not be constructed.

To support triple-key partitioning for links that open in a new tab,
this CL allows HistoryTabHelper to populate top_level_url with the URL
of the LiveOriginalOpenerChain in the case that
PreviousPrimaryMainFrameURL and Opener are empty or invalid.

For simplicity, this CL also combines/simplifies the existing code
paths for replacing PreviousPrimaryMainFrameURL with Opener (that added
support for context clicks). Now, all of this code lives in
HistoryTabHelper.

Bug: 400828290
Change-Id: Ie39210ff6ebacd327e430d19ae6df269c4e750d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6388415
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437866}