Move ExtensionService::CreateUpdateInstaller() to ExtensionUpdater
ExtensionService has many dependencies that aren't yet supported on
desktop Android, so we're moving code out of it into places where we
can port. ExtensionService::CreateUpdateInstaller() has a single
caller, in ExtensionUpdater, so move it there.

The ExtensionUpdater tests extensively use a delegate interface with
the CreateUpdateInstaller() method for sensing. Rename that delegate
interface to CrxInstallerProviderForTest and make it only used in
tests. Production code can directly call CreateUpdateInstaller().

Provide a test-only constructor that maps to existing unit test
code for ExtensionUpdater, to minimize churn in the tests.

Some tests that use an "empty" extension service are updated to
have an ExtensionUpdater instance so they can call the method.
This will be simplified in a future CL when ExtensionUpdater
becomes a profile keyed service.

No functional changes, just refactor.

Bug: 404943906
Change-Id: Ic263f9b7b2869e3bb8ac9a16bedc2c6965037227
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387852
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437890}