[Blink] Re-enable IsPaintableChecksResourceProviderInsteadOfBridge
I disabled the IsPaintableChecksResourceProviderInsteadOfBridge
killswitch in [1] defensively due to a realization that
it interacts negatively with the resource provider recreation code in
CanvasRenderingContext2D::Restore() - that code recreates the resource
provider without recreating the bridge, which with this killswitch
enabled can violate the invariant
that the bridge (and hibernation handler) are always present if there
is a valid resource provider present. This CL eliminates that
problematic interaction by having
HTMLCanvasElement::DiscardResourceProvider() not drop the bridge if this
killswitch enabled but simply clear hibernation instead.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6357235

Bug: 40280152, 399731315
Change-Id: Ifa57270f076fdbb1521248402bfa577de0f46b5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6375246
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1438007}