Add test to verify cache with GetCookiesOnSet feature enabled
A previous CL introduced functionality to update the cookie jar
cache on set: http://crrev.com/c/6048964

This CL adds a test to verify that the cache is invalidated if
devtools restriction is toggled between operations. This is done
simply by getting a cookie with and without devtools enabled to
verify a Get->Get flow. We also set a cookie without devtools, turn
on devtools, then get the cookie to verify the Set->Get flow.

Bug: 380864710
Change-Id: I2798037562a44ef957fdced655011240f53a61a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6386640
Reviewed-by: Camille Lamy <clamy@chromium.org>
Commit-Queue: Joshua Thomas <masnoble@chromium.org>
Reviewed-by: Aldo Culquicondor <acondor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1438084}