[Android][Signin] Attempt workaround for class cast exception
As an attempt of working around the cast exception in the sign-in
activity, this CL sets a classloader to use for parsing parcelable
intents before retrieving any extra from the extras Bundle.

Since the retrieval makes the Bundle unparcel all extras, the right
ClassLoader needs to be set before that to avoid classloader mismatch.

Bug: 394559360
Change-Id: I1fa0c40a54b093f0dc627da170613fd3738e4c17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6382098
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Menghan Yang <myuu@google.com>
Cr-Commit-Position: refs/heads/main@{#1438096}