[Test Automation] Add timeout context when waiting for a state change
This reverts commit 5ffefe4466c0bd8fca7b72e7fe639108f544aca0. The new
functionality was added as part of another CL but turned out to be
buggy due to an order-of-operations issue on test teardown.

This new functionality is useful because on timeout when waiting for
a state change, it will print out the last check it did of that state,
using MatchAndExplain(). This means that you can (usually) see what the
actual value was, and why it did not match the value or matcher that
the step was waiting for.

This can be extremely useful in debugging tests that fail on a
WaitForState() step.

Original change's description:
> [gardener] Partial revert of mac hover fix change.
>
> https://crrev.com/c/6373791 introduced some changes to
> StateObserverElementT that appear to trigger some MSAN
> use-of-uninitialized-value failures for several tests.
>
> We can't revert the whole commit because other tests would fail, but
> this should mitigate the failures for the time being pending a
> follow-up fix.
>
> Bug: 406190898
> Change-Id: I6a1dd7699f66bb3f5f035f59c4623c9c05c17fe0
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6393998
> Owners-Override: Justin Lulejian <jlulejian@chromium.org>
> Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
> Reviewed-by: Dana Fried <dfried@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1437794}

Bug: 406190898
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Change-Id: I79cc5d48f6e39dd70a55d2b0baf3ee01a95fb161
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6395341
Reviewed-by: Justin Lulejian <jlulejian@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1438098}