cdm: Attach crash key when CDM process has debugger attached
The CDM process should detect and add a crash key when crashes occur in
the utility process when a debugger is attached. The detection of the
debugger attached should happen before the CDM process is sandboxed due
to intricacies with the debugger detection code.

We also add a histogram to track when a debugger is attached.

testing:

For testing purposes, I also added a statement to log to histogram when
a debugger is not detected. I removed that in our main code because we
don't expect a debugger to be attached the majority of the time.

1. When a debugger is not attached, the histogram reports false for the
debugger attached.

2. When I just called `gdb --args ./out/Default/chrome`, this just
attaches a debugger to the browser process. In this case, the histogram
also reports false.

3. When I call `.\tools\win\windbg64.bat -G -g -o
.\out\Default\chrome.exe --enable-logging --user-data-dir=~/test/`,
which attaches a debugger to all child processes, the histogram reports
true! and the CDM utility process also crashes. Which is what we want to
detect.

Bug: 394910957
Change-Id: I59795d59779cda835ca23ebc7e3e9a42152c2166
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6378344
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Reviewed-by: Feras Aldahlawi <feras@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1438099}